requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.1.0
langsmith>=0.1.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
docker>=7.0.0
rich
tavily-python
langgraph-cli>=0.1.0